requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.8
django-bootstrap-v5==1.0.11
django-filter==23.5
python-dotenv==1.0.0
soupsieve==2.5
